refactor: clean up dead code for env API
