+++ packages/astro/src/core/config/index.ts
@@ -3,7 +3,6 @@ export {
 	resolveConfigPath,
 	resolveRoot,
 } from './config.js';
-export { createConsoleLogger, createNodeLogger } from './logging.js';
 export { mergeConfig } from './merge.js';
 export { createSettings } from './settings.js';
 export { loadTSConfig, updateTSConfigForFramework } from './tsconfig.js';
