+++ packages/astro/src/core/config/logging.ts
@@ -1,5 +1,4 @@
 import type { AstroInlineConfig } from '../../types/public/config.js';
-import { consoleLogDestination } from '../logger/console.js';
 import { Logger } from '../logger/core.js';
 import { nodeLogDestination } from '../logger/node.js';
 
@@ -9,12 +8,5 @@ export function createNodeLogger(inlineConfig: AstroInlineConfig): Logger {
 	return new Logger({
 		dest: nodeLogDestination,
 		level: inlineConfig.logLevel ?? 'info',
-	});
-}
-
-export function createConsoleLogger(level: AstroInlineConfig['logLevel']): Logger {
-	return new Logger({
-		dest: consoleLogDestination,
-		level: level ?? 'info',
 	});
 }
